Moved ResolveLibrary function in the correct source file
